refactor transaction
